+++ laktory/enums.py
@@ -4,6 +4,9 @@ from typing import Any
 
 import narwhals as nw
 
+from laktory._settings import settings
+from laktory.typing import AnyFrame
+
 
 class DataFrameBackends(str, Enum):
     def _generate_next_value_(name, start, count, last_values):
@@ -34,6 +37,21 @@ class DataFrameBackends(str, Enum):
     #
     # UNKNOWN = auto()
     # """Unknown backend."""
+
+    @classmethod
+    def from_any(cls, backend: str | nw.Implementation | AnyFrame = None):
+        if backend is None:
+            return DataFrameBackends(settings.dataframe_backend)
+        elif isinstance(backend, str):
+            return DataFrameBackends(backend)
+        elif isinstance(backend, nw.Implementation):
+            return DataFrameBackends.from_nw_implementation(backend)
+        elif isinstance(backend, AnyFrame):
+            return DataFrameBackends.from_df(backend)
+        elif isinstance(backend, DataFrameBackends):
+            return backend
+        else:
+            raise NotImplementedError()
 
     @classmethod
     def from_nw_implementation(