+++ laktory/models/dataframe/dataframeschema.py
@@ -7,6 +7,7 @@ from pydantic import Field
 from pydantic import model_validator
 
 from laktory._logger import get_logger
+from laktory.enums import DataFrameBackends
 from laktory.models.basemodel import BaseModel
 from laktory.models.dataframe.dataframecolumn import DataFrameColumn
 from laktory.models.dtypes import DType
@@ -55,13 +56,42 @@ class DataFrameSchema(BaseModel):
 
         return data
 
-    # Narwhals
+    # -------------------------------------------------------------------------------- #
+    # Class Methods                                                                    #
+    # -------------------------------------------------------------------------------- #
+
+    # From Narwhals
+    @classmethod
+    def from_narwhals(cls, schema: nw.Schema) -> "DataFrameSchema":
+        """Create a DataFrameSchema from a Narwhals schema"""
+        columns = [
+            DataFrameColumn(name=name, dtype=DType.from_narwhals(dtype))
+            for name, dtype in schema.items()
+        ]
+        return cls(columns=columns)
+
+    # -------------------------------------------------------------------------------- #
+    # Instance Methods                                                                 #
+    # -------------------------------------------------------------------------------- #
+
     def to_narwhals(self) -> nw.Schema:
         """Returns a Narwhals schema object"""
         cols = {}
         for c in self.columns:
             cols[c.name] = c.dtype.to_narwhals()
         return nw.Schema(cols)
+
+    def to_native(
+        self, dataframe_backend: nw.Implementation | DataFrameBackends | None = None
+    ):
+        dataframe_backend = DataFrameBackends.from_any(dataframe_backend)
+
+        if dataframe_backend == DataFrameBackends.PYSPARK:
+            return self.to_pyspark()
+        elif dataframe_backend == DataFrameBackends.POLARS:
+            return self.to_polars()
+        else:
+            raise NotImplementedError()
 
     # Polars
     def to_polars(self):
@@ -74,16 +104,20 @@ class DataFrameSchema(BaseModel):
         return pl.Schema(cols)
 
     # Spark
-    def to_spark(self):
+    def to_pyspark(self):
         """Returns a Spark schema object"""
         import pyspark.sql.types as T
 
         columns = []
         for c in self.columns:
-            _type = c.dtype.to_spark()
+            _type = c.dtype.to_pyspark()
             columns += [T.StructField(c.name, _type, c.nullable)]
 
         return T.StructType(columns)
+
+    # -------------------------------------------------------------------------------- #
+    # Outputs                                                                          #
+    # -------------------------------------------------------------------------------- #
 
     # String
     def to_string(self, indent=None):