+++ tests/dataframe/test_dataframeschema.py
@@ -2,7 +2,9 @@ import narwhals as nw
 import polars as pl
 import pyspark.sql.types as T
 
+from laktory.enums import DataFrameBackends
 from laktory.models import DataFrameSchema
+from laktory.models import DType
 from laktory.models import dtypes
 
 s = DataFrameSchema(
@@ -13,6 +15,18 @@ s = DataFrameSchema(
         {"name": "vals", "dtype": {"name": "list", "inner": "string"}},
     ]
 )
+
+
+def test_dtype_round_trip():
+    nw_dtype = nw.dtypes.Int64()
+    dtype = DType.from_narwhals(nw_dtype)
+    assert dtype.to_narwhals() == nw.dtypes.Int64()
+
+    nw_dtype = nw.List(
+        inner=nw.dtypes.Struct({"x": nw.dtypes.Float64(), "id": nw.dtypes.String()})
+    )
+    dtype = DType.from_narwhals(nw_dtype)
+    assert dtype.to_narwhals() == nw_dtype
 
 
 def test_validation():
@@ -40,7 +54,7 @@ def test_narwhals():
 
 
 def test_spark():
-    assert s.to_spark() == T.StructType(
+    assert s.to_pyspark() == T.StructType(
         [
             T.StructField("x", T.LongType(), True),
             T.StructField("y", T.DoubleType(), True),
@@ -56,8 +70,19 @@ def test_polars():
     )
 
 
+def test_native():
+    assert s.to_native(DataFrameBackends.POLARS) == pl.Schema(
+        {"x": pl.Int64, "y": pl.Float64, "s": pl.String, "vals": pl.List(pl.String)}
+    )
+
+
 def test_to_string():
     assert (
         s.to_string()
         == '{"x": "Int64", "y": "Float64", "s": "String", "vals": "List(String)"}'
     )
+
+
+def test_from_narwhals():
+    schema = DataFrameSchema.from_narwhals(s.to_narwhals())
+    assert schema == s